# deploy/nginx/search-gateway-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VRChat Feedback Search Gateway",
    "version": "1.1.0",
    "summary": "Public search API. Agents use the short /api/agent routes; the website uses Algolia-compatible InstantSearch.",
...
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Same-origin; server URL **`/`**."
    }
  ],
  "security": [],
...
  "paths": {
    "/api/search": {
      "post": {
        "operationId": "searchMultiQuery",
        "summary": "Run one or more InstantSearch queries.",
        "description": "Accepts a JSON **`array`** of **InstantSearch** multi-query payloads. Supply **`mode`** as an HTTP **`query`** parameter (never inside the request **`body`**); **`mode=lucene`** enables Lucene syntax for each **`params.query`**. Maximum **`body`** size **512 KiB**.",
        "tags": [
          "Search"
        ],
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "lucene"
              ]
            },
            "description": "Parsing mode for each item's **`params.query`**. Omit for default matching. **`lucene`** selects Lucene syntax; see overview."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MultiQueryRequest"
              },
              "examples": {
                "minimalQuery": {
                  "summary": "Single query, equivalent to **GET** **`/api/search?q=ui&hitsPerPage=15&page=0`**",
                  "value": [
                    {
                      "indexName": "feedback-posts",
                      "params": {
                        "query": "ui",
                        "hitsPerPage": 15,
                        "page": 0
                      }
                    }
                  ]
                },
                "facetedQuery": {
                  "summary": "Faceted query with **`facetFilters`** and **`indexName`** **`feedback-posts_score_desc`**",
                  "value": [
                    {
                      "indexName": "feedback-posts_score_desc",
                      "params": {
                        "query": "desktop crash",
                        "hitsPerPage": 20,
                        "page": 0,
                        "facets": [
                          "status",
                          "board_slug"
                        ],
                        "facetFilters": [
                          [
                            "status:open"
                          ]
                        ]
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "**`SearchResponse`** **`results`** entries with **`hits`**, paging, **`facets`**, and metadata (**`nbHits`**, **`nbPages`**, **`processingTimeMS`**, etc.).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "**`body`** is not a JSON **`array`**, or search expression is **`Invalid search query`** (bad Lucene/filter syntax).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "malformedBody": {
                    "summary": "Expected **`InstantSearch`** **`multi-query`** **`array`**",
                    "value": {
                      "message": "Expected a JSON array of InstantSearch multiple-queries payloads"
                    }
                  },
                  "invalidSearch": {
                    "summary": "Invalid Lucene/filter (**`Invalid search query`**)",
                    "value": {
                      "message": "Invalid search query",
                      "detail": "Search service rejected the request (HTTP 400): ..."
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "**`request`** **`body`** exceeds **512 KiB**. The response may omit **`application/json`** and use **`text/plain`**."
          },
          "500": {
            "description": "Search operation could not be completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "search failed"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "searchSingleQueryOrDiscovery",
        "summary": "Run a single query via query string, or return a discovery document.",
        "description": "Runs a streamlined single-query search when any of **`q`**, **`query`**, **`hitsPerPage`**, or **`page`** appears; otherwise returns a discovery document (**`DiscoveryResponse`**).\n\n### Single-query search\n\nProviding any query parameter noted above invokes one search against **`feedback-posts`**, equivalent to a one-element **`POST`** payload. **`?mode=lucene`** applies Lucene parsing to the text query consistently with **`POST`**.\n\n| **`Parameter`** | Equivalent **`InstantSearch`** field | **`Default`** | **`Constraints`** |\n| ----- | ----- | ----- | ----- |\n| **`q`** or **`query`** | **`params.query`** | `\"\"` | **`q`** overrides **`query`** |\n| **`hitsPerPage`** | **`params.hitsPerPage`** | **`50`** | Clamped to **`[1, 500]`**; invalid inputs use **`50`** |\n| **`page`** | **`params.page`** | **`0`** | Clamped to **`[0, 9999]`**; invalid inputs use **`0`** |\n| **`mode`** | Parsing profile for **`params.query`** | _(omit ⇒ default)_ | **`lucene`** only; ineffective without **`q`**, **`query`**, **`hitsPerPage`**, or **`page`** above |\n\nAdvanced behaviours—multiple simultaneous queries, faceting definitions, granular filters beyond these shortcuts, choice of sort replica—require **`POST`**.\n\n### Discovery\n\nAbsent all of **`q`**, **`query`**, **`hitsPerPage`**, and **`page`** (even if **`mode=lucene`** is present), **`GET`** returns endpoint metadata rather than hits.",
        "tags": [
          "Search"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Search terms. Maps to **`params.query`**. Takes precedence over **`query`** when both are supplied.",
            "example": "desktop crash"
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Alias of **`q`**. Ignored when **`q`** is also present."
          },
          {
            "name": "hitsPerPage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 500
            },
            "description": "Page size. Defaults to **`50`**. Values outside **`[1, 500]`** or non-numeric input fall back to **`50`**."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0,
              "maximum": 9999
            },
            "description": "Zero-based page index. Defaults to **`0`**. Values outside **`[0, 9999]`** or non-numeric input fall back to **`0`**."
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "lucene"
              ]
            },
            "description": "Interpretation mode for the URL **`q`**/**`query`** search terms (**`params.query`** after mapping). Omit for default matching. **`lucene`** enables Lucene syntax; **`mode`** alone never runs a search—pair it with **`q`**, **`query`**, **`hitsPerPage`**, or **`page`**."
          }
        ],
        "responses": {
          "200": {
            "description": "Either search results (**`SearchResponse`**) when search parameters were included, or a discovery document (**`DiscoveryResponse`**).",
            "headers": {
              "Link": {
                "$ref": "#/components/headers/ServiceLink"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/SearchResponse"
                    },
                    {
                      "$ref": "#/components/schemas/DiscoveryResponse"
                    }
                  ]
                },
                "examples": {
                  "discovery": {
                    "summary": "Discovery (**`GET`** missing **`q`**, **`query`**, **`hitsPerPage`**, **`page`**)",
                    "value": {
                      "endpoints": {
                        "POST": "/api/search",
                        "GET": "/api/search?q=terms&hitsPerPage=50&page=0",
                        "agentSearch": "/api/agent/search?q=terms&limit=10&page=0",
                        "agentPost": "/api/agent/posts/{board}/{urlName}",
                        "index": "/api/index",
                        "openapi": "/openapi.json"
                      },
                      "description": "Agents: **`GET /api/agent/search`** and **`GET /api/agent/posts/{board}/{urlName}`**. UI: **`POST`**/**`GET /api/search`**. Contract: **`openapi.json`**."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "invalid search expression (**`Invalid search query`**).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Invalid search query",
                  "detail": "failed to parse ..."
                }
              }
            }
          },
          "500": {
            "description": "Search failed. Not used for **`DiscoveryResponse`**-only **`GET`** calls.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "search failed"
                }
              }
            }
          }
        }
      },
      "options": {
        "operationId": "searchCorsPreflight",
        "summary": "CORS preflight.",
        "description": "Cross-origin resource sharing (CORS) preflight. Returns **`204`** with permissive **`Access-Control-*`** headers for browser clients.",
        "tags": [
          "Search"
        ],
        "responses": {
          "204": {
            "description": "**`No`** **`Content`**."
          }
        }
      }
    },
    "/api/index": {
      "get": {
        "operationId": "indexGeneration",
        "summary": "Current OpenSearch backing index name.",
        "description": "Returns the backing index currently published behind the **`feedback-posts`** alias. The name changes after each ingest alias swap, so clients can poll this endpoint and re-run the current search when **`index`** changes. Response is not cached.",
        "tags": [
          "Operations"
        ],
        "responses": {
          "200": {
            "description": "Current backing index name.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexResponse"
                },
                "example": {
                  "index": "feedback-posts-202608130112"
                }
              }
            }
          },
          "500": {
            "description": "Backing index lookup failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "index lookup failed"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/search": {
      "get": {
        "operationId": "agentSearch",
        "summary": "Search posts for an agent.",
        "description": "Short read-only search. Each hit is a title, canonical post **`url`**, board, status, score, comment count, created time, and a **`details`** snippet of at most **200** characters. **`_source`** omits voters, comments, and the full body.\n\n**`q`** uses the same strict text match as the website default (not Lucene). A non-empty **`q`** orders hits by relevance, then newest **`created`**. An empty or omitted **`q`** lists newest posts. **`board`** and **`status`** are exact filters on **`board.urlName`** and **`status`**.\n\n**`limit`** defaults to **10** and is clamped to **`[1, 20]`**; invalid input uses **10**. **`page`** is zero-based and defaults to **0**; invalid input uses **0**. **`page * limit`** must be below **50000**.",
        "tags": [
          "Agent"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Search text. Omit or send an empty string to list newest posts."
          },
          {
            "name": "board",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Exact **`board.urlName`**, for example **`feature-requests`**."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Exact post **`status`**, for example **`open`**."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 20,
              "default": 10
            },
            "description": "Page size. Defaults to **10**. Values above **20** are clamped to **20**. Invalid or non-positive input uses **10**."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "Zero-based page index. Defaults to **0**."
          }
        ],
        "responses": {
          "200": {
            "description": "Short hit list.",
            "headers": {
              "Link": {
                "$ref": "#/components/headers/ServiceLink"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSearchResponse"
                },
                "example": {
                  "q": "avatar performance",
                  "board": "feature-requests",
                  "status": "open",
                  "page": 0,
                  "limit": 10,
                  "nbHits": 1,
                  "hits": [
                    {
                      "post_id": "5aff34492d6f472658e02dfa",
                      "title": "Support for VRM",
                      "url": "https://feedback.vrchat.com/feature-requests/p/support-for-vrm",
                      "board": "feature-requests",
                      "status": "open",
                      "score": 40,
                      "commentCount": 3,
                      "created": "2018-05-19T00:00:00.000Z",
                      "snippet": "Quest avatar limits…"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameter, or **`page`** is past the result window.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "invalid board"
                }
              }
            }
          },
          "500": {
            "description": "Search failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "search failed"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/posts/{board}/{urlName}": {
      "get": {
        "operationId": "agentPost",
        "summary": "Fetch one post for an agent.",
        "description": "One post identified by exact **`board.urlName`** and **`urlName`**. **`details`** is truncated to **4000** characters (**`detailsTruncated`**). **`comments`** keeps at most **8** public comments, pinned first and then newest, each **`value`** truncated to **500** characters. Deleted, internal, private, spam, and blank comments are omitted. Voters are omitted.",
        "tags": [
          "Agent"
        ],
        "parameters": [
          {
            "name": "board",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Exact **`board.urlName`**."
          },
          {
            "name": "urlName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Exact post **`urlName`**."
          }
        ],
        "responses": {
          "200": {
            "description": "The post, truncated for an agent context window.",
            "headers": {
              "Link": {
                "$ref": "#/components/headers/ServiceLink"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPost"
                },
                "example": {
                  "post_id": "5aff34492d6f472658e02dfa",
                  "title": "Support for VRM",
                  "url": "https://feedback.vrchat.com/feature-requests/p/support-for-vrm",
                  "board": "feature-requests",
                  "status": "open",
                  "score": 40,
                  "commentCount": 12,
                  "created": "2018-05-19T00:00:00.000Z",
                  "details": "Please add VRM import.",
                  "detailsTruncated": false,
                  "comments": [
                    {
                      "author": "Ada",
                      "created": "2020-01-01T00:00:00.000Z",
                      "pinned": true,
                      "value": "Staff note.",
                      "valueTruncated": false
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "No indexed post matches that board and url name.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "post not found"
                }
              }
            }
          },
          "500": {
            "description": "Lookup failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "search failed"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "headers": {
      "ServiceLink": {
        "description": "Machine contract (**`rel=service-desc`**) and human API reference (**`rel=service-doc`**).",
        "schema": {
          "type": "string",
          "example": "</openapi.json>; rel=\"service-desc\"; type=\"application/openapi+json\", </openapi.html>; rel=\"service-doc\""
        }
      }
    },
    "schemas": {
      "MultiQueryRequest": {
        "type": "array",
        "description": "InstantSearch **`array`**; each **`object`** has **`indexName`** + **`params`**. Allowed **`params`** keys follow Algolia conventions; **`params.query`** parsing depends on URL **`mode`** (**`lucene`** or omitted — see overview).",
        "minItems": 1,
        "items": {
          "$ref": "#/components/schemas/MultiQueryRequestItem"
        }
      },
...
      "SearchResponse": {
        "type": "object",
        "description": "Algolia-style search-response envelope listing one **`results`** entry per submitted query.",
        "required": [
          "results"
        ],
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            }
          }
        },
        "additionalProperties": true
      },
...
      "IndexResponse": {
        "type": "object",
        "description": "Backing index currently published behind the **`feedback-posts`** alias.",
        "required": [
          "index"
        ],
        "properties": {
          "index": {
            "type": "string",
            "description": "OpenSearch backing index name (timestamped, e.g. **`feedback-posts-202608130112`**)."
          }
        },
        "additionalProperties": false
      },
      "DiscoveryResponse": {
        "type": "object",
        "description": "Discovery payload from **`GET /api/search`** when none of **`q`**, **`query`**, **`hitsPerPage`**, or **`page`** are present; **`mode`** alone never performs a search.",
        "required": [
          "endpoints",
          "description"
        ],
        "properties": {
          "endpoints": {
            "type": "object",
            "required": [
              "POST",
              "GET",
              "agentSearch",
              "agentPost",
              "index",
              "openapi"
            ],
            "properties": {
              "POST": {
                "type": "string",
                "description": "**`POST`** **`/api/search`** URL."
              },
              "GET": {
                "type": "string",
                "description": "**`GET`** example (**`/api/search`** URL **`query`** string)."
              },
              "agentSearch": {
                "type": "string",
                "description": "**`GET /api/agent/search`** URL."
              },
              "agentPost": {
                "type": "string",
                "description": "**`GET /api/agent/posts/{board}/{urlName}`** URL."
              },
              "index": {
                "type": "string",
                "description": "**`GET`** **`/api/index`** URL (current backing index name)."
              },
              "openapi": {
                "type": "string",
                "description": "**`openapi`** key: **`/openapi.json`** path **`string`**."
              }
            },
            "additionalProperties": false
          },
          "description": {
            "type": "string",
            "description": "Synthesized **`discovery`** narration (**Markdown** tolerated)."
          }
        },
        "additionalProperties": false
      },
...
      "AgentSearchResponse": {
        "type": "object",
        "required": [
          "q",
          "page",
          "limit",
          "nbHits",
          "hits"
        ],
        "properties": {
          "q": {
            "type": "string"
          },
          "board": {
            "type": "string",
            "description": "Present when the request filtered by **`board`**."
          },
          "status": {
            "type": "string",
            "description": "Present when the request filtered by **`status`**."
          },
          "page": {
            "type": "integer",
            "minimum": 0
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 20
          },
          "nbHits": {
            "type": "integer",
            "minimum": 0
          },
          "hits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentHit"
            }
          }
        },
        "additionalProperties": false
      },
      "AgentComment": {
        "type": "object",
        "required": [
          "pinned",
          "value",
          "valueTruncated"
        ],
        "properties": {
          "author": {
            "type": "string"
          },
          "created": {
            "type": "string"
          },
          "pinned": {
            "type": "boolean"
          },
          "value": {
            "type": "string",
            "description": "Comment text, at most **500** characters."
          },
          "valueTruncated": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "AgentPost": {
        "type": "object",
        "description": "One post. **`comments`** contains at most **8** comments.",
        "required": [
          "comments"
        ],
        "properties": {
          "post_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "description": "Canonical post URL: **`https://feedback.vrchat.com/{board}/p/{urlName}`**."
          },
          "board": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "score": {
            "type": "integer"
          },
          "commentCount": {
            "type": "integer",
            "description": "Full comment count on the post, which can be larger than **`comments.length`**."
          },
          "created": {
            "type": "string"
          },
          "details": {
            "type": "string",
            "description": "Post body, at most **4000** characters."
          },
          "detailsTruncated": {
            "type": "boolean"
          },
          "comments": {
            "type": "array",
            "maxItems": 8,
            "items": {
              "$ref": "#/components/schemas/AgentComment"
            }
          }
        },
        "additionalProperties": false
      },
      "Error": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "**`message`**: **`string`**; **`detail`**: supplemental **`diagnostics`** (optional)."
          },
          "detail": {
            "type": "string",
            "description": "Supplemental diagnostics when **`message`** is **`Invalid search query`**; omit when **`body`** is not an **`InstantSearch`** **`array`**, etc."
          }
        },
        "additionalProperties": false
      }
    }
  }
}